Add ns timer

+++ firmware/src/drivers/counter.h
@@ -18,9 +18,6 @@
 #define COUNTER_H
 
 
-#include <stdint.h>
-
-
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -29,7 +26,7 @@ extern "C" {
 void counter_init (void);
 void counter_task (void);
 
-uint32_t counter_freq_hz (void);  // return of zero is no frequency available
+double counter_freq_hz (void);  // return of zero is no frequency available
 
 
 #ifdef __cplusplus
